# src/main/resources/openapi/openapi.yml
openapi: 3.0.3
info:
  title: Stock Service Microservice
  version: 1.0.0
  description: Stock APIs for managing Stock.
servers:
  - url: http://localhost:8080/api/v1
paths:
...
  /auth/register:
    post:
      summary: Register a new user
      tags:
        - ServiceAuth
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegistrationRequest'
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Request contains incorrect syntax or cannot be processed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 400
                title: "invalid-request"
                detail: "Invalid values sent in the request"
        '409':
          description: Username already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 409
                title: "conflict"
                detail: "Username already exists"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - code
        - title
        - detail
      properties:
        code:
          description: "A number that represents a type of errors, grouping those of similar characteristics"
          type: integer
        title:
          description: "A identifier for the error code"
          type: string
        detail:
          description: "A description of the error"
          type: string
...
    UserRegistrationRequest:
      type: object
      properties:
        username:
          type: string
          description: The unique username of the user
          minLength: 4
          maxLength: 50
          example: johndoe
        password:
          type: string
          description: The password of the user
          example: P@ssw0rd
        roles:
          type: array
          items:
            type: string
          description: The roles assigned to the user
          example: [ "USER" ]
      required:
        - username
        - password
    UserResponse:
      type: object
      properties:
        username:
          type: string
          description: The unique username of the user
          example: johndoe
        roles:
          type: array
          items:
            type: string
          description: The roles assigned to the user
          example: [ "USER" ]